+++ content-scripts/shared/skills.js
@@ -1,0 +1,18 @@
+export const skillIndexOutfielders = {
+    "SC": 0,
+    "OP": 1,
+    "BC": 2,
+    "PA": 3,
+    "AE": 4,
+    "CO": 5,
+    "TA": 6,
+    "DP": 7
+}
+
+export const skillIndexGoalkeepers = {
+    "RE": 0,
+    "GP": 1,
+    "IN": 2,
+    "CT": 3,
+    "OR": 4
+}