+++ content-scripts/pages/player/player+training.js
@@ -9,10 +9,4 @@ export async function processPlayerTraining() {
         console.debug("Player training table not found, skipping...")
         return
     }
-
-    const potentialRiseIndicators = playerTrainingTable.querySelectorAll(`tbody > tr > td > span.text-info`)
-    const potentialRiseAdditionalClass = utils.pluginNodeClass + "PotentialRiseAdditionalClass"
-    for (const indicator of potentialRiseIndicators) {
-        indicator.classList.add(potentialRiseAdditionalClass)
-    }
 }