+++ content-scripts/release_notes.md
@@ -54,3 +54,5 @@
 # 6.3.0
 - Moved the injuries table to the right side and updated the styles
 - Fixed the transfer list footer to top relocation
+- Reverted the color change of the potential rise indicator on the `/player` Training tab since it was fixed in the game
+- Implemented special talents modifiers on the `/player` page ([#55](https://github.com/lawicko/final-whistle-assistant/issues/55))